requirements.txt
Flask
tensorflow-cpu==2.15.0
numpy
Pillow
gunicorn
